Fix parameter for custom_access_token_expires_in
`custom_access_token_expires_in` parameter could be an instance of:

* `Doorkeeper::OAuth::PreAuthorization`
* `Doorkeeper::OAuth::Client`
* `Doorkeeper::Application`

`Doorkeeper::OAuth::Client` is a small wrapper around
`Doorkeeper::Application`. `PreAuthorization` is an entirely different
object. This patch checks if we are about to send a `PreAuthorization`,
and if so we grab the associated client. This way, the method always
receives an Oauth Client as parameter.

[fixes #617]

Conflicts:
	lib/doorkeeper/oauth/authorization/token.rb
	lib/doorkeeper/oauth/client.rb